+++ .prettierrc.js
@@ -1,5 +1,14 @@
 'use strict';
 
 module.exports = {
+  plugins: ['prettier-plugin-ember-template-tag'],
+  overrides: [
+    {
+      files: '*.{js,ts,gjs,gts}',
+      options: {
+        singleQuote: true,
+      },
+    },
+  ],
   singleQuote: true,
 };